Edit. 팀페이지 이슈 부분 해결
