+++ src/components/teamPage/chat/MessageBox.jsx
@@ -12,11 +12,11 @@ const MessageBox = ({ t }) => {
   return (
     <MessageSection>
       <MessageImageBox>
-        <MessageImage src={t.profileImg} alt="" />
+        <MessageImage src={t?.profileImg} alt="" />
       </MessageImageBox>
       <MessageTextBox>
-        <MessageNickName>{t.nickName}</MessageNickName>
-        <MessageContent>{t.comment}</MessageContent>
+        <MessageNickName>{t?.nickName}</MessageNickName>
+        <MessageContent>{t?.comment}</MessageContent>
       </MessageTextBox>
     </MessageSection>
   );
