+++ src/pages/joinLogin/SignUp.jsx
@@ -110,6 +110,7 @@ const SignUp = () => {
             email: email,
             profileImg: 'https://imhannah.me/common/img/default_profile.png',
             bookmarks: [],
+            teamID: [],
           });
 
         updateProfile(authService.currentUser, {
