+++ src/pages/mate/MateWrite.jsx
@@ -143,35 +143,23 @@ const MateWrite = () => {
               partyLocation,
               partyTime,
             },
-          })
-            .then(() => {
-              updateDoc(doc(db, 'user', authService.currentUser.uid), {
-                teamID: [..._teamID, teamID],
-              });
-            })
-            .then(() => {
-              setDoc(doc(db, 'teamChat', teamID), {
-                teamID,
-                // message: [
-                //   {
-                //     comment: '',
-                //     uid: authService.currentUser.uid,
-                //     profileImg: profileUserInfo,
-                //     nickName:''
-                //   },
-                // ],
-              });
-            })
-            .catch(() => {
-              console.log('채팅 에러');
-            })
-            .catch(() => {
-              console.log('user 에러');
-            });
+          });
+        })
+        .then(() => {
+          updateDoc(doc(db, 'user', authService.currentUser.uid), {
+            teamID: [teamID],
+          });
+        })
+        .then(() => {
+          setDoc(doc(db, 'teamChat', teamID), {
+            teamID,
+            message: [],
+          });
         })
         .catch(() => {
-          console.log('team 에러');
+          console.log('팀페이지 에러');
         });
+
       console.log('업로드 성공');
       navigate(`/mate`);
     } catch (error) {
@@ -183,7 +171,6 @@ const MateWrite = () => {
     if (!currentUser) return;
     getUserInfo();
     getLeaderImg();
-    console.log(currentUser);
   }, [currentUser]);
 
   return (
